Add an in-memory pubsub system.
I think this will be useful for a few things initially, e.g.:

  * Updates on what is currently playing and the current position in the
    track, e.g., to show in the main UI.
  * Progress messages when scanning the database.

Later on, I think it will also be useful for things like automatically
adding things to the playlist when the player is near the end of the
playlist. Or submitting played track information to last.fm. Or anything
else where the producer of the information really doesn't care what the
consumer does with the information.