Try to set the current duration from STREAM_START.
It seems that sometimes STREAM_START is too early and the duration isn't
ready yet, but other times the first _poll_status() call after
STREAM_START is too late and the duration is already the duration for
the following stream. Hopefully this change will make the duration
correct as much of the time as possible.