Add error policy for a new play order module.
Play orders will determine the previous and next playable units. An
order will be able to stop either because it has nothing to do (e.g., at
the end of a playlist) or because of an error. Consumers of orders
probably won't care about the distinction, but implementations of orders
will care when composing themselves with other order implementations.

This commit adds an exception, a policy enum, and a decorator to
implement the policy. Order implementations will come in follow-up
commits.